+++ NEWS.md
@@ -1,3 +1,9 @@
+# TesiproV 0.9.2.0
+
+Better implementation of parallel computing in the unix and macOS environment.
+Changed RNG in parallel approach.
+
+
 # TesiproV 0.9.1.0
 
 * First release.